Move cursor context manager into BaseCog
